Add files via upload
This dataset comprises national time series for representing building heat pumps in power system models. The heat demand of buildings and the coefficient of performance (COP) of heat pumps is calculated for 16 European countries from 2008 to 2018 in an hourly resolution. Heat demand time series for space and water heating are computed by combining gas standard load profiles with spatial temperature and wind speed reanalysis data as well as population geodata. The profiles are year-wise scaled to 1 TWh each.

https://data.open-power-system-data.org/when2heat/2019-08-06